backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database (document metadata only)
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Vector Store
chromadb

# Settings
pydantic-settings==2.7.0
pydantic==2.10.4
python-dotenv==1.0.1

# OCR
pytesseract==0.3.13
Pillow==11.1.0

# Image Preprocessing
opencv-python-headless==4.10.0.84
scikit-image==0.25.0
numpy==2.2.2

# PDF Generation & Processing
reportlab==4.2.5
pypdf==5.1.0
pdf2image==1.17.0

# Embeddings - Cohere
cohere==5.13.6

# LLM - Google Gemini
google-generativeai==0.8.3

# Utilities
aiofiles==24.1.0
httpx==0.28.1

# Authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.2.0
pandas